+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: vockers <vockers@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/02 14:42:28 by vockers       #+#    #+#                 */
-/*   Updated: 2023/10/03 14:44:09 by vockers       ########   odam.nl         */
+/*   Updated: 2023/10/03 16:33:13 by vockers       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,10 @@ char	*ft_strnstr(const char *big, const char *little, size_t len)
 	i = 0;
 	if (*little == '\0')
 		return ((char *)big);
-	while (big[i] && i < len)
+	while (big[i] && i < (int)len)
 	{
 		j = 0;
-		while (big[i] == little[j] && i < len)
+		while (big[i] == little[j] && i < (int)len)
 		{
 			j++;
 			i++;
