+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: vockers <vockers@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/10/02 14:42:15 by vockers       #+#    #+#                 */
-/*   Updated: 2023/10/03 14:55:39 by vockers       ########   odam.nl         */
+/*   Updated: 2023/10/03 16:33:21 by vockers       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	if (!sub)
 		return (NULL);
 	i = 0;
-	while (s[start + i] && i < len)
+	while (s[start + i] && i < (int)len)
 	{
 		sub[i] = s[start + i];
 		i++;
